requirements.txt
alembic==1.8.1
anyio==3.6.1
Authlib==1.0.1
backports.cached-property==1.0.2
certifi==2022.6.15
cffi==1.15.1
click==8.1.3
cryptography==37.0.4
ecdsa==0.18.0
fastapi==0.79.0
FastAPI-SQLAlchemy==0.2.1
fastapi-sso==0.6.2
graphql-core==3.2.1
h11==0.12.0
httpcore==0.15.0
httptools==0.4.0
httpx==0.23.0
idna==3.3
itsdangerous==2.1.2
jose==1.0.0
Mako==1.2.1
MarkupSafe==2.1.1
oauthlib==3.2.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
Pygments==2.12.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
python3-pip-autoremove==0.9.1
PyYAML==6.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
strawberry-graphql==0.121.0
tenacity==8.0.1
typing_extensions==4.3.0
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.16.0
websockets==10.3
